ipa: raspberrypi: AWB: Ignore invalid statistics zones correctly
Regions of the image where the statistics are not trusted should not
be put in the zones_ list.

This also means that the AWB may fail to run initially, so
async_results_ needs to be initialised to some kind of default.

Finally, the RGB zone structures can be zero-initialised safely now,
the previous value of -1 would seem unusual.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Reviewed-by: Naushir Patuck <naush@raspberrypi.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>