ipa: raspberrypi: Add support for imx327-based SE327M12 module
This patch adds support for the Soho Enterprises SE327M12 module.  The
sensor is an imx327 which therefore uses the imx290 kernel driver and
CamHelper.

To use this module and camera tuning, place the following in the
/boot/config.txt file:

dtoverlay=imx290,clock-frequency=37125000

and then the existing imx290.json file must be over-written with
se327m12.json

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>